requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.4.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.7.0
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
flake8==7.1.1
fsspec==2025.3.0
huggingface-hub==0.29.3
idna==3.10
kombu==5.4.2
mccabe==0.7.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sqlparse==0.5.3
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
